requirements.txt
email_validator
flask
Flask-FlatPages
Flask-Reuploaded
Flask-SimpleMDE
Flask-SQLAlchemy
Flask-WTF
Pillow
python-dotenv
sendgrid
